Update user.cfc
I found that wrapping a serialized Json object in a struct puts '\' characters all over the Json object so I found that bypassing serializeJson and hand coding the JSON worked.  And it is consistent with other parts of your code.

This also I need Authorization variable to appear in the Header box in the RestPlaygound but I will work on that later.